# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Emumet Account Service API",
    "description": "Account Service for ShuttlePub",
    "contact": {
      "name": "turtton"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/accounts/{account_id}/metadata": {
      "post": {
        "tags": [
          "Metadata"
        ],
        "description": "Create a metadata entry for the specified account.",
        "operationId": "create_metadata",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Metadata created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/accounts/{account_id}/metadata/{id}": {
      "put": {
        "tags": [
          "Metadata"
        ],
        "description": "Update a metadata entry.",
        "operationId": "update_metadata",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Metadata nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Metadata updated"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Metadata"
        ],
        "description": "Delete a metadata entry.",
        "operationId": "delete_metadata",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Metadata nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Metadata deleted"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/accounts/{account_id}/profile": {
      "put": {
        "tags": [
          "Profile"
        ],
        "description": "Update the profile of the specified account.",
        "operationId": "update_profile",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Profile updated"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Profile"
        ],
        "description": "Create a profile for the specified account.",
        "operationId": "create_profile",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Account nanoid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/metadata": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "description": "Retrieve metadata entries for the given account IDs.",
        "operationId": "get_metadata_batch",
        "parameters": [
          {
            "name": "account_ids",
            "in": "query",
            "description": "Comma-separated account IDs",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/profiles": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Retrieve profiles for the given account IDs.",
        "operationId": "get_profiles_batch",
        "parameters": [
          {
            "name": "account_ids",
            "in": "query",
            "description": "Comma-separated account IDs",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateMetadataRequest": {
        "type": "object",
        "required": [
          "label",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "CreateProfileRequest": {
        "type": "object",
        "properties": {
          "banner": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "MetadataResponse": {
        "type": "object",
        "required": [
          "account_id",
          "nanoid",
          "label",
          "content"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "nanoid": {
            "type": "string"
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "required": [
          "account_id",
          "nanoid"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "banner_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "nanoid": {
            "type": "string"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateMetadataRequest": {
        "type": "object",
        "required": [
          "label",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "banner": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}